+++ src/main/client/system/index.ts
@@ -7,7 +7,9 @@ import './raycasts.js';
 import './screenshot.js';
 import './serverConfig.js';
 import './serverKeybind.js';
+import './serverKeypress.js';
 import './stats.js';
+import './vehicle.js';
 import './waypoint.js';
 
 alt.log('Loaded System');